refactor(#742): relocate namespace static dispatch to call-namespace-static.ts
Wave B chunk 3, step 2. Move compileNamespaceStaticCall (Symbol / Reflect /
Promise / JSON / Date statics) out of calls.ts into a new sibling module
src/codegen/expressions/call-namespace-static.ts. Nine calls.ts internals
exported to it; calls.ts imports compileNamespaceStaticCall back (lazy cycle,
same pattern as call-identifier.ts / call-builtin-static.ts).

calls.ts: 14,484 -> 12,564 LOC. compileCallExpression: ~8,410 -> ~6,480 LOC.
New call-namespace-static.ts: 2,028 LOC.

Pure relocation — emitted Wasm byte-identical (prove-emit-identity IDENTICAL,
39/39). tsc 0. Smoke test extended with Symbol.for / Date.UTC cases. loc-budget
+ coercion-sites (one relocated __is_truthy site, net-zero) pass with allowances
in the #742 frontmatter; oracle-ratchet net-zero.

✓
